dashboardv1 file added and everything in dashboard cleared out
